requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
app @ file:///C:/Users/ThinkPad/Desktop/New%20folder%20%282%29/test/app-0.1.0-py3-none-any.whl#sha256=302d8f547d26d116fca72ec58f6dbbb851e506e5c42e48d5d5116f2b4ad50d2b
APScheduler==3.11.0
argcomplete==3.6.2
attrs==25.4.0
backoff==2.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
confluent-kafka==2.12.0
decorator==5.2.1
dnspython==2.8.0
dotenv==0.9.9
fastapi==0.121.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
loguru==0.7.3
motor==3.7.1
multidict==6.7.0
packaging==25.0
pipx==1.7.1
platformdirs==4.3.8
playwright==1.54.0
propcache==0.4.1
psycopg2==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
scrapfly-sdk==0.8.23
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.3
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.38.0
win32_setctime==1.2.0
yarl==1.22.0
